Avoid constructing Allocator instances on test module load
Google Benchmark constructs instances of all registered benchmarks in a
global static variable, and stores pointers to them in a vector. Because of
this, the destruction order can cause the application to crash at exit.
This avoids that problem by using pointers to allocator instances in the
benchmarks, instead of storing them by value.

Signed-off-by: Chris Burel <burelc@amazon.com>