Fix flaky test - increase timeout
When I measured this locally on a machine without load, it was
something near 100ms to perform this test.

When I added a stress load to the machine, the time it takes
was anywhere between 100ms and 8 seconds.

Added comment indicating on the choice.  Also fixes a rare crash
in Qt if the thread is deleted before it stops.

Signed-off-by: Nicholas Lawson <70027408+lawsonamzn@users.noreply.github.com>