Fixing full screen shadow bug with msaa. (#14342)
This PR fixes the full screen shadow bug when msaa is turned on and also adds support for shader variants on full screen triangle passes.

The core issue was that full screen shadows assume the depth pixel they are reading is at the pixel center, but when msaa is turned on, that's not the case. The depth resolve simply takes the first sample position, so the sample being read by full screen shadows comes from the location of the first sample. This location varies depending on if msaa 2x, 4x, 8x, or 16x is used, but thankfully msaa sample positions seem to be standardized across all platforms.

The solution was to just offset the pixel position by its real sample location, and not assume center-of-pixel. This was done using a shader option for maximum performance, which allows some of the math to be baked at compile time. Shader variants were provided for no msaa, msaa 2x, and msaa 4x. 8x and 16x will just use the root variant, but those modes are rarely used because of the memory requirements.

Support needed to be added to the FullScreenTriangle pass in order for FullScreenShadows to use shader options. To not duplicate work, I moved the pass's custom pipeline state handling to RPI::PipelineStateForDraw. That class takes care of most of the work with shader variants, so it allowed FullScreenTrianglePass to not grow significantly.