Material Canvas: Sorting generated material type groups and properties
This change implements automatic sorting of material type groups and properties emitted from material canvas.

Material input nodes are used to expose material properties on the generated material type. These nodes have fields that allow the user to configure the material type property names, groups, and descriptions. The material type groups and properties will be registered in the order nodes are traversed. That order is not deterministic. There is currently no variable manager or other mechanism to explicitly dictate the organization of groups and properties. In lieu of that, the groups and properties will be sorted alphabetically.

Resolves https://github.com/o3de/o3de/issues/14263

Resolves ATOM-18318

Signed-off-by: gadams3 <guthadam@amazon.com>