Fixed error while loading PhysX materials. (#14462)
Signed-off-by: moraaar <moraaar@amazon.com>

## What does this PR do?

Fixes a bug that causes a crash when while waiting for a physics material asset to load the asset manager dispatches events for other assets. These events can cause `OnAssetReady` of other assets to be called and if those end up trying to create the same physics material then it leads to the crash.

Solved by creating the material first, then block until loaded.

Also renamed the utility function `GetShapesFromAsset` to `CreateShapesFromAsset` since that reflects better what it does.

## How was this PR tested?

Run PhysX unit tests.
Tested using a local level that was crashing while loading due to the issue described above. After this fix the level loads correctly. Also debugged the loading of colliders and rigid bodies to verify it works correctly.