fix empty path defaulting to current working dir
any function that used QString_To_Py_Path() and provided an empty string was getting converted to the current working directory in Python instead of the Python None type, e.g. when getting all gems an empty path is provided, but to Python that looks like you're passing the current working directory as a project path - which throws a warning.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>