Fix compile issues and also found a way to avoid removing constexpr
Thanks to @lumberyard-employee-dm for the hints on how to avoid
having to change the constexpr.  This results in fewer risky changes
to the code.

The "default" constexpr hash function is implemented as
```
return static_cast<size_t>(source type);
```

This means that instead of specializing the hash template for your
given type you can just provide an operator size_t() on your type.

This means that you can put your hash function inside your type and
not necessarily cause the bus to instantiate in the header, until used.

Signed-off-by: Nicholas Lawson <70027408+lawsonamzn@users.noreply.github.com>