Updated the `AZ_COMPONENT_BASE` macro to no longer rely on RTTI.
The CreateDescriptor functions defined by the macro now uses AzTypeInfo
to query the a componetns Uuid and type name.

This allows the `AZ_RTTI_NO_TYPE_INFO_IMPL` macro to be use privately
within a gem for a component without needing the `RTTI_Type` and
`Rtti_TypeName` functions for linkage.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>