Replaced usage of the azrtti_typeid macro in SerializeContext.h
The macro usage has been replaced with calls to `AzTypeInfo<T>::Uuid()`
which removes the dependency on the `RTTI_Type` function needing to be
implemented within a Gem MODULE target.

This allows a Gems' AZ::Module derived class to reflect it's component
descriptors to the SerializeContext without needing an implementation of
an `RTTI_Type` static function for the component types.
The implementation will not accessible for a Component within the Gems MODULE target if the component uses the `AZ_RTTI_NO_TYPE_INFO_IMPL` macro.

In this case the implementation is usually only a symbol with the Gems
private STATIC library target.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>