Update SettingsRegistry project path determination logic
The project-path command line parameter s no longer converted into a
--regset parameter, instead the FindProjectRoot function directly checks
the `--project-path` command line parameter first.

If the --project-path command line parameter is not set, then the first
ancestor path of the executable directory that contains a project.json
file is used as the project path.

Finally if the previous 2 checks of the `--project-path` command line
fails and scanning upwards for a projec.json fails, the
`/Amazon/AzCore/Bootstrap/project_path` setting is read to find the
project path.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>