fix include function stomping, ignore old engines
The functions defined in VersionUtils were getting stomped because of how cmake deals with functions that try to use the same name.
Moving the function logic into o3deConfigVersion.cmake solves the issue
Also, remove logic that lets a user register an upgraded project with an old engine that does no version compatibility checks.
Users can revert to the old EngineFinder.cmake from the old engine or create a project with the old engine if they want to do this.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>