Reduced recompile time when changing O3DE_STACK_CAPTURE_DEPTH
CMake contains a cache variable called `O3DE_STACK_CAPTURE_DEPTH` which
determines the amount of stack frames to capture when a memory leak
occurs in the C++ UnitTest.

When that cache variable is changed, it modified the public
COMPILE_DEFINITION of `O3DE_STACK_CAPTURE_DEPTH` on the `O3DEKernel`
library target.

As the `O3DEKernel` is a dependency of AzCore and AzCore is a dependency
on all the other libraries within the code base, this required the
entire code base to recompile when the option is changed.

To reduce the time needed to recompile, the `O3DE_STACK_CAPTURE_DEPTH`
CMake cache variable will only change the COMPILE_DEFINITION for the
`OSAllocator.cpp`, `PoolAllocator.cpp` and `SystemAllocator.cpp`.

Therefore only those 3 files need to recompile.
There is still a cost to relinking AzCore to other libraries, but is a
lot less than recompiling every source file.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>