This is the new ShaderVariantListBuilder. It fixes the problem we have today that any change to a `.shadervariantlist` file would cause recompilation of all the shader variants listed in the `.shadervariantlist`.
This builder works based on the concept of
Intermediate Assets.

---
=== The first time around ===
The ShaderVariantListBuilder will find
that a new `<shader_name>.shadervariantlist` file has been added to the filesystem
and it will proceed to create a new Intermediate Source Asset
called `<shader_name>.hashedvariantlist` which will contain
the "hashed" version of the same variants found in the original
`<shader_name>.shadervariantlist`. Each variant will also be flagged as **IsNew=true**.

The ShaderVariantAssetBuilder will wake up and find that there's a list
where all variants are marked as "IsNew=true"and will proceed to compile
all those variants.

---

=== The second time around ===
After the user modifies `<shader_name>.shadervariantlist`,
either by adding a new variant or modifying an existing one, the ShaderVariantListBuilder
will open the previously generated `<shader_name>.hashedvariantlist` and compare
the hashes with the new list of variants. Only those variants that are found to be new
,or their hash has changed since the last time, will be marked as **IsNew=true**, all
other variants will be marked as **IsNew=false**. The results will overwrite the
Intermediate `<shader_name>.hashedvariantlist` file.

The ShaderVariantAssetBuilder will wake up and find that the  `<shader_name>.hashedvariantlist`
contains some variants marked as **IsNew=false**, and those won't be recompiled.
The other variant(s) found as **IsNew=true** will be compiled.

---

Other goodies in this PR:
Added a couple of Registry Key properties that relate to the ShaderVariantListBuilder:
And they have been set by default in atom_shaders.setreg:
```json
{
    "O3DE": {
        "Atom": {
            "Shaders": {
                "ShaderVariantListBuilder" : {
                    "EnableHashCompare": false, // Setting this to false will always rebuild all variants when a .shadervariantlist changes.
                    "SuddenChangeInMinutes": 5, // If a .shadervariantlist changes several times within this time period, the IsNew status of each previous variant in .shadervariantlist will be preserved.
                }
            }
        }
    }
}
```

---
REMARK1: This PR introduces the ShaderVariantListBuilder,
but for the sake of simplifying the Code Review, this builder
will be inactive because it listens for file with extension `*.shadervariantlist2`
which no such file exist yet in the engine.

REMARK2: During debugging the ShaderVariantListBuilder the new algorithms
found that `o3de\Gems\Atom\Feature\Common\Assets\Shaders\SkyAtmosphere\SkyRayMarching.shadervariantlist`
had a redundant variant with StableId=4, which was the same variant as StableId=2.
This has been fixed by removing the variant with StableId=4.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>