Added declaration and definition macros for AzTypeInfo
These macros allows the definitions for the GetO3deTypeName and
GetO3deTypeId overloads to be implemented in a translation unit to
reduce build time compiling inline functions across multiple translation
units.

Removed constexpr static member from the AzTypeInfo template.
This reduces compile time and allows the AzTypeInfo template to be
externed and explicitly instantiated in a cpp file.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>