Replaced static member GetO3deTypeName/GetO3deTypeId functions with friend functions
This allows providing a consistent call pattern for invoking the
GetO3deTypeName and GetO3deTypeId and allows removal of a template
instantiation for each call of the function.
A template function was required before to defer the check of the
whether the member versions of GetO3deTypeName/GetO3deTypeId or unqualified versions could be called until instantiation time.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>