Updating the BehaviorClass unwrapper to copy its callable
The callable is copied into a `unique_ptr<void>` with a custom deleter.
This allows member function pointers on non-Windows platforms to be used
with the unwapper.
Furthermore this functor objects to be used to unwrap a class as well.

Additional changes: Update the BehaviorClassUnwrapperFunction to supply
a `BehaviorObject&` which is populated with the unwrapped object, typeid
and rtti object.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>