Fix mipmap eviction. (#14882)
If the set of loaded images have varying numbers of mipmaps, the previous logic wouldn't evict as many mipmaps as it could, because it would stop at the first failure case. For example, if 10 textures have 1 mipmap and 2 textures have 5 mipmaps, it will evict one mipmap from each of those 2 textures, put them at the front of the list, and then stop evicting because it will check a texture with 1 mipmap. With this change, it will keep searching and evict 4 mipmaps from those 2 textures, freeing more memory before failing.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>