SerializeContext header include dependency reduction
The goal of changes in this commit is to remove unneccesary inclusion of SerializeContext.h from header files.

Whenever possible if a Reflect function was defined in header, it's
definition was moved to a cpp file and the header was replaced with a
forward declaration to ReflectContext.

The SerializeContext VersionConverter functions that were members of a class declaration were
instead moved to cpp files as static functions.
Declaring the SerializeContext VersionConverter functions in header
inhibits forward declaration of the SerializeContext due to the need to
access the local class of SerializeContext::DataElementNode

The AnimSplineTrack specializations for Vec2 for both the LyShine and
Maestro Gems were moved to a cpp file, with forward declarations for
the `TAnimSplineTrack<Vec2>` and `TUiAnimSplineTrack<Vec2>` member
functions added to remove the include of the
AnimSplineTrack_Vec2Specialization.cpp file

Finally the `AZ_TYPE_INFO_SPECIALIZE*` macros were moved from TypeInfo.h
to TypeInfoSimple.h which allows code to avoid including `TypeInfo.h`
when needing to use that macro.

The TypeInfo.h include specializes AZ TypeInfo for AZStd container
templates such as `AZStd::vector`, `AZStd::unordered_map`, etc...
So this should save a little bit of include time.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>