Makes the visuals of physics joint scale as you zoom
Previously, if you zoom in far enough the physics joint visuals would
cover the entire screen unlike the rest of the gizmos and visualizations.

This also adds a different visualizer for the revolving joint (the
hinge joint with no limits).  I made sure that adding rotation in the
positive direction caused the body to rotate in the direction the
arrows indicate.

Signed-off-by: Nicholas Lawson <70027408+lawsonamzn@users.noreply.github.com>