Fixed assert in the ActionManagerSystemComponent::Init
When the ActionManagerSystemComponent is loaded in an Application that
hasn't create a QGUIApplication, it asserts when a QWidget is created
with the message
> QWidget: Must construct a QApplication before a QWidget

The fix is to avoid creating the QWidget if there is no QApplication.

Side Note: We should investigate why the ActionManagerSystemComponent is creating a
QWidget. It is the only SystemComponent that creates a QT GUI Object.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>