Removed unused private member variables
The `-Wunused-private-field` warning is now triggering when using the clang compiler, so the unused private members have been removed.

There is two cases where the members are only used with the MSVC compiler and those variables have had the warning surpressed using [[maybe_unused]]

The ScriptCanvasDeveloperGem `EditorKeyActions.h` and `EditorMouseActions.h` are the cases where the private members are actually used.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>