Atom Tools: Updating asset browser, selectors, and enumeration to ignore the cache and intermediate asset folders
Now that intermediate assets are being used to generate material types and shader permutations, the files are showing up in the asset browser and asset selection controls, adding a lot of noise and confusion because of the extra options.

This PR hides them from the asset browser and asset selection controls in material editor, material canvas, and other tools based on atom tools framework.

A custom, general purpose asset browser entry filter was also added that allows using the standard function instead of creating complex combinations of other filter types.

Signed-off-by: gadams3 <guthadam@amazon.com>