Fixes sparse vector to be able to handle non-trivial destructors. (#15014)
Previously it was possible for objects stored in SparseVector or MultiSparseVector to have their destructors called multiple times. This happened because the underlying vector(s) have no knowledge of which objects stored within have already been destroyed and represent empty slots. To fix this, the classes detect if any of the underlying types use non-trivial destructors, and if so, go through the free slots and default-construct objects so they can be safely destroyed later. Generally, SparseVector will only be used with simple data structs, and rarely anything with a non-trivial destructor. In cases where it is using a type with non-trivial destructor, the additional cost will only be on shutdown and when Clear() is called. However, it's worth keeping in mind if the default construction cost is high, this may not be a good container type to use.

SparseVector is only used with a non-trivial destructor in one place, which happens to be harmless. In SkyAtmosphere there is a struct with a destructor that just resets the values on the struct and isn't really necessary.

Signed-off-by: Ken Pruiksma <pruiksma@amazon.com>