Fix Physics Material Asset property from showing invalid asset in components (#15019)
## What does this PR do?

Physics material reflection to edit context does not need an asset assigned, the components handle that if there is no asset it uses the default material. In order to show the text "(default)" when there is no asset assigned Physics materials were passing an null asset to the DefaultAsset property. This was ok before, but after [this change](https://github.com/o3de/o3de/pull/13959) this is reported as an error.

- Fixed by passing the default physics material from the material manager, which is an in-memory asset. 
- This was still reported as error because the asset property handler expected an asset job for the default asset, but since it's an in-memory asset there is no job for it and it was reported it as missing asset. Fixed by only reporting as missing asset if it is not found in asset manager.

## How was this PR tested?

Checked PhysX and terrain components that use physics material property and verified that they don't show as invalid anymore when it doesn't have a material assigned.

Signed-off-by: moraaar <moraaar@amazon.com>