Fixes race condition when debugging Lua files. Closes #14823
The TcpConnection class has a m_sendRingBuffer
that is NOT thread safe.

NetworkSystemComponent uses the Main thread
to send messages, and hence use the m_sendRingBuffer.

On the other hand, before this fix, the RemoteToolsSystemComponent.cpp
was sending packets with the help of a second thread controlled
by the class RemoteToolsOutboxThread.cpp. This second thread is totally
unnecessary and even worst was causing race conditions. This race
condition was corrupting the messages being sent from Editor.exe
to LuaIDE.exe.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>