Split the Parallel_Thread.Test unit test into smaller UnitTest
This is being done to better debug failures in the future
that occurs on Linux tracked in issue #14536

Fixed the AZStd::thread creation calls to properly supply the thread
descriptor argument as the first argument to `AZStd::thread`

Fixed intermittent memory leak in Parallel thread unit test due to the
thread swap unit test detaching the thread and not joining.

This created a race condition where a thread can still be running at the
end of the Parallel_Thread.Test when the memory leak checker is checking
for existing allocations.

fixes #14536

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>