Updated the ShaderVariantAssetBuilder.
This new version doesn't consume *.shadervariantlist files.
Instead it consumes *.hashedvariantlist and *.hashedvariantinfo files.

As a reminder, the ShaderVariantListBuilder consumes *.shadervariantlist and
produces Intermediate Assets: One *.hashedvarianlist + Several *.hashedvariantinfo

In turn, this new ShaderVariantAssetBuilder has two inputs:
- First Input: *.hashedvariantlist. Output: ShaderVariantTreeAsset (*.azshadervarianttree)
- Second Input: *.hashedvariantinfo. Ouputs: ShaderVariantAsset (*.azshadervariant)
  one for each combination of RHI+Supervariant+StableId.

**What's the benefit**:
The main benefit is that whenever the original *.shadervariantlist file changes,
let's say because the user added a new variant, or because the user modified an existing
variant, only the new/updated variant will be rebuilt.
This will have a huge impact in the future where there can be, let's say 1000 variants
listed in the *.shadervariantlist file. If you add +1 variant, for a new total of 1001 variants,
then only the new ShaderVariantAsset will be compiled, instead of recompiling all 1001 ShaderVariantAssets.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>