Fixed an issue where a serialization version number went down instead of up, which was causing issues with upgrading and error reporting.
Some assets were emitting errors that they were saved with a future version, because the version went down from 14 to 2 when the previous version was added. This fixes the version, makes it so the node version changing causes ScriptCanvas files to re-process, and I bumped the prefab builder version because it was emiting the same error and wasn't processing when script canvas files changed, due to a gap in the job dependency graph, which I will investigate soon and cut a ticket if one is needed.

Signed-off-by: AMZN-stankowi <4838196+AMZN-stankowi@users.noreply.github.com>