Fixes https://github.com/o3de/o3de-extras/issues/118
You cannot use FD_SET and select() on file handle numbers greater
than 1023, as the fd set structure is actually a bit mask and has
a fixed size.  Using file handles greater than 1023 will cause it to
buffer overflow.

This fixes the problem in 2 ways
1. It reduces the default number of keep-open handles to a more
   reasonable number (128) across the board on all platforms, as it is
   generally not neccessary for O3DE applications to keep 1024 files
   open all the time for any real reason.
2. It uses epoll instead of select and FD_SET to wait for data to be
   ready, future-proofing this on linux in case it occurs again or
   for whatever reason we use a lot of file handles again.

Note that the file handle count max 1024 is a global system wide
problem, so any time any application uses a large number of file handles
other applications may die, if they use FD_SET in their code.

This is why the action taken is to both future proof O3DE applications
but ALSO reduce the amount of handles the applications themselves
claim so that other programs are not affected.

Signed-off-by: Nicholas Lawson <70027408+lawsonamzn@users.noreply.github.com>