Adding a CompressionLZ4 sub gem to the Compression Gem
The CompressionLZ4 gem wraps the lz4 3rdParty library and implements the
Compression Gem ICompressionInterface and IDecompressionInterface.

A lz4 compressor implementation has been added that registers with the
the CompressionRegistrar which is available in tools code.
In client code a lz4 decompressor implementation has been added that
registers with the DecompressionRegistrar.

Added UnitTest to the Compression Gem to validate the
CompressionRegistrar and DecompressionRegistrar works properly.

Updated the Compression Interface with a CompressBound function which
provides an upper bound on potential compressed data size given the
uncompressed data size.

Also added more detailed error result structure which can provide an
error string on errors that occurs during compression/decompression

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>