Fixes windows deadlock
It turns out windows deadlocks if a child process writes enough data
to both stderr (unbuffered) and stdout (buffered) because it is incorrectly
using the WaitForMultipleObjects api on objects it cannot wait on (pipes).

This causes WaitForMultipleObjects to return immediately with 0.
This causes it to believe that the stdout handle is ready for reading.
This causes it to call ReadFile on stdout handle and block forever.

This means that it was always busy looping before, the busy loop is just
super obvious now with comments.
I leave it up to future self or others to refactor this code to use
asynch io functions and named pipes instead of WaitForMultipleObjects API.

Signed-off-by: Nicholas Lawson <70027408+lawsonamzn@users.noreply.github.com>