Fix error when prefab generator skips generation
When a prefab group exists in the manifest but generation is skipped due to a registry setting, it leaves behind an empty prefab group which causes exporting to fail later on since the group still exists.  This fix causes the group to be removed before skipping the entry and adds an info message to notify users that a registry setting caused generation to be skipped.

Signed-off-by: amzn-mike <80125227+amzn-mike@users.noreply.github.com>