[Terrain] Ray tracing terrain sectors (#15331)
This PR removes the old rather prototypey ray tracing support in terrain and replaces it with ray traced geometry that should exactly match the geometry in the forward pass. This is done by copying the geometry data for each sector into formats that are digestible by ray tracing. Color is still not supported, so all of the terrain is assigned a base color of medium gray (0.18).