Fix deadlock in FileDescriptorCapturer
The FileDescriptorCapturer class would deadlock when the file descriptor data
that was being captured filled up the pipe buffer used to redirect the
output.

This would result in the application being blocked attempting to write
captured output to one end of the pipe on the same thread where a read of the other end of the pipe would occur later.

The changes here involve making a dedicated "Flush" thread, whose purposes
to perform reads on one of the pipe when it receives data and then forward that data to a user provided callback.

On Windows a Completion Routine is used to perform asynchronous IO the
flush thread to have it yield when it is waiting for input on the read
side of the pipe

On Linux and Android, `epoll` is used to wake up the flush thread when
data is available to read from the pipe

On MacOS and iOS, `kqueue` is used to receive a signal to read data from
the pipe.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>