Adds the ability to generate UVs automatically during mesh import
This adds a new modifier to the available list of modifiers you can
add to a mesh during its import in its import settings.

The new modifier lets you select how to generate UVs for a mesh.
Currently, only 1 generator is supported (Spherical Positional
projection) but all the hard work here is done to support additional
unwrappers in the future by following the pattern.

Signed-off-by: Nicholas Lawson <70027408+lawsonamzn@users.noreply.github.com>