Added "OrderOnly" job dependency.
This new job dependency is helpful when it is needed
that "Asset A" completes before "Asset B" runs,
but don't re-run "Asset B" just because "Asset A" ran.

The ShaderVariantAssetBuilder is the first builder that
takes advantange of this new job dependency because it needs
the ShaderVariantTreeAsset to complete before any ShaderVariantAsset,
But ShaderVariantAssets do not rebuild simply because
the ShaderVariantTreeAsset was rebuilt.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>