Remove m_buildTimeStamp from ShaderAsset and ShaderVariantAsset.
It was necessary a few years ago, but with the recent
improvements in the Asset System notifications it is
not necessary anymore and removing it also adds stability
to Asset Bundling. The main reason problem of a having
a build time stamp is that the same shader build in different dates
would appear as different assets, but now, regardless of the date
of compilation the Shader assets remain the same, assuming the
code does not change.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>