fix incorrect model used in update
refresh() was telling the inspector to update using the proxy selection model which does not always map to the source model, resulting in failed model updates.
Switching to itemFromIndex() makes the issue obvious because it causes asserts when the index is invalid instead of quietly generating a new (invalid) index.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>