Adding support for SystemFile API to support wraping Std handles
The `stdin`, `stdout` and `stderr` file handles are wrapped by static
functions on the SystemFile class that allows using the SystemFile class
to read from stdin or write to stdout/stderr using its API.

The Std handles are setup to not close on destruction of the SystemFile
class by default.

So it is safe to store these SystemFiles instances by value.

Finally a "skip close on destruction" flag has been added to the SystemFile open
mode flags to allow an open SystemFile to skip closing an open file
handle in the SystemFile destructor.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>