Optimize initial AP startup time
Added a string path cache to the FileStateCache to avoid re-normalizing and lowercasing the same string many times.  Profiling showed this as a hot spot and it was being called over 5M times during startup.
Added a dependency cache to QueryAbsolutePathDependenciesRecursive.  Profiling also showed this as a hotspot for several different items.  The cache only applies during startup and is cleared/disabled after reaching idle state.
Added an optional return of the scanfolder for FindFirstMatchingFile to optimize one of the SourceAssetReference calls during the above.

Signed-off-by: amzn-mike <80125227+amzn-mike@users.noreply.github.com>