Avoid using the `ClassBuilder` after it has been destroyed
`behaviorContext->Class<LuaMaterialFunctorAPI::RuntimeContext>`
returns a `ClassBuilder<LuaMaterialFunctorAPI::RuntimeContext>` by
value. That object is is then passed to ClassBuilder::operator->(),
which transforms it into a pointer. Then it is passed to a bunch of
functions that all return a `ClassBuilder*`, which is why they can chain
together to make `ClassBuilder::Method()` calls. By the end of that
expression on line 320, there is no owner of the `ClassBuilder` value, and
its destructor runs. However, a pointer to it is saved in the builder
variable declared on line 320. That invalid pointer is then passed to
`LuaMaterialFunctorAPI::ReadMaterialPropertyValues::ReflectSubclass`, and
undefined behavior results.

This retains the object returned by `BehaviorContext::Class()` until the
end of the enclosing function.

Signed-off-by: Chris Burel <burelc@amazon.com>