Fix off-by-one error when allocating XML string
This code uses `rapidxml::allocate_string` to allocate enough space for a
string. The code afterwards explicitly null-terminates the string, so it
tries to allocate enough space for the current length of the string + 1.
However, the `allocate_string` call also copies the source string into
place, copying the same number of bytes as the `size` argument. Because we
asked for one more than the current size of the string, this code ends up
in an invalid read of size 1.

This fixes that bug by doing the copy ourselves. We allocate enough space
for the null terminating byte, then copy just the source string (whose size
does not include space for the null terminator), then finally explicitly
null terminate it.

Signed-off-by: Chris Burel <burelc@amazon.com>